Rename secrets to secret_values to avoid numpy error: https://discuss.python.org/t/importerror-cannot-import-name-randbits/16919/5
