+++ plugins/together/newhelm/suts/together_client.py
@@ -6,7 +6,7 @@ from together.utils import response_status_exception  # type: ignore
 from newhelm.prompt import ChatPrompt, ChatRole, SUTOptions, TextPrompt
 from newhelm.prompt_formatting import format_chat
 from newhelm.record_init import record_init
-from newhelm.secrets import SecretValues, UseSecret
+from newhelm.secret_values import SecretValues, UseSecret
 from newhelm.secrets_registry import SECRETS
 from newhelm.sut import PromptResponseSUT, SUTCompletion, SUTResponse
